Renaming for clarity and to work around a bug in nose: https://github.com/nose-devs/nose/issues/24
