+++ chime/instance_setup/user-data.sh
